# src/test/resources/swagger/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OGC API - Processes",
    "description": "Example API Definition for OGC API - Processes",
    "contact": {
      "name": "Open Geospatial Consortium",
      "email": "info@ogc.org"
    },
    "license": {
      "name": "OGC License",
      "url": "http://www.opengeospatial.org/legal/"
    },
    "version": "0.1"
  },
  "servers": [
    {
      "url": "http://localhost/testogcapi",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "retrieve the list of jobs.",
        "description": "Lists available jobs.  For more information, see [Section 12](https://docs.ogc.org/is/18-062r2/18-062r2.html#Job_list). ",
        "operationId": "getJobs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "processID",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "accepted",
                  "running",
                  "successful",
                  "failed",
                  "dismissed"
                ]
              }
            }
          },
          {
            "name": "minDuration",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "maxDuration",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "datetime",
            "in": "query",
            "description": "Either a date-time or an interval, half-bounded or bounded. Date and time expressions adhere to RFC 3339. Half-bounded intervals are expressed using double-dots.  Examples:  * A date-time: \"2018-02-12T23:20:50Z\" * A bounded interval: \"2018-02-12T00:00:00Z/2018-03-18T12:31:12Z\" * Half-bounded intervals: \"2018-02-12T00:00:00Z/..\" or \"../2018-03-18T12:31:12Z\"  Only features that have a temporal property that intersects the value of `datetime` are selected.  If a feature has multiple temporal properties, it is the decision of the server whether only a single temporal property is used to determine the extent or all relevant temporal properties.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of jobs for this process.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{jobId}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "retrieve the status of a job",
        "description": "Shows the status of a job.   For more information, see [Section 7.10](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_retrieve_status_info). ",
        "operationId": "getStatus",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "local identifier of a job",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The status of a job.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "500": {
            "description": "A server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "cancel a job execution, remove a finished job",
        "description": "Cancel a job execution and remove it from the jobs list.  For more information, see [Section 14]https://docs.ogc.org/is/18-062r2/18-062r2.html#Dismiss). ",
        "operationId": "dismiss",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "local identifier of a job",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The status of a job.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "500": {
            "description": "A server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{jobId}/results": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "retrieve the result(s) of a job",
        "description": "Lists available results of a job. In case of a failure, lists exceptions instead.  For more information, see [Section 7.11](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_retrieve_job_results). ",
        "operationId": "getResult",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "local identifier of a job",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The processing results of a job.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InlineOrRefData"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "500": {
            "description": "A server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{jobId}/results/{outputId}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "retrieve the result of a job",
        "description": "Shows a single output of a job. In case of a failure, shows exceptions instead.  For more information, see [Section 7.11](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_retrieve_job_results). ",
        "operationId": "getOutputResult",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "local identifier of a job",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outputId",
            "in": "path",
            "description": "identifier of the output of a job",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The processing output result of a job.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              },
              "application/geo+json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              },
              "application/geo+json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "500": {
            "description": "A server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              },
              "application/geo+json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "exception": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        },
        "description": "JSON schema for exceptions based on RFC 7807"
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Metadata": {
        "type": "object"
      },
...
      "StatusInfo": {
        "required": [
          "id",
          "status",
          "type"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Metadata"
            }
          },
          "id": {
            "type": "string"
          },
          "processID": {
            "type": "string",
            "format": "uri"
          },
          "type": {
            "type": "string",
            "example": "[process, wps, openeo]"
          },
          "request": {
            "$ref": "#/components/schemas/StatusInfoAllOfRequest"
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "running",
              "successful",
              "failed",
              "dismissed"
            ]
          },
          "message": {
            "type": "string"
          },
          "exception": {
            "$ref": "#/components/schemas/exception"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "started": {
            "type": "string",
            "format": "date-time"
          },
          "finished": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "progress": {
            "maximum": 100,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          }
        }
      },
      "StatusInfoAllOfRequest": {
        "type": "object"
      },
...
      "JobList": {
        "required": [
          "jobs",
          "links"
        ],
        "type": "object",
        "properties": {
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatusInfo"
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          }
        }
      },
      "InlineOrRefData": {
        "type": "object"
      },
...
    }
  }
}